requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.8.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==3.0.3
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.13.0
MarkupSafe==2.1.5
multidict==6.1.0
packaging==25.0
propcache==0.2.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
yarl==1.15.2
zipp==3.20.2
